+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added Vow cards
 - Added Dmu cards
+- Added possible card variation screen
+- Added color and type filter
 
 ### Changed